+++ tasmota/tasmota_xdrv_driver/xdrv_22_sonoff_ifan.ino
@@ -244,7 +244,7 @@ void SonoffIfanUpdate(void)
  * Interface
 \*********************************************************************************************/
 
-bool Xdrv22(uint8_t function)
+bool Xdrv22(uint32_t function)
 {
   bool result = false;
