+++ tasmota/tasmota_xdrv_driver/xdrv_44_miel_hvac.ino
@@ -1285,7 +1285,7 @@ static void (*const miel_hvac_cmnds[])(void) PROGMEM = {
 #endif
 };
 
-bool Xdrv44(uint8_t function) {
+bool Xdrv44(uint32_t function) {
 	bool result = false;
 	struct miel_hvac_softc *sc = miel_hvac_sc;
