+++ tasmota/tasmota_xdrv_driver/xdrv_82_esp32_ethernet.ino
@@ -297,7 +297,7 @@ void CmndEthSetIpConfig(void) {
  * Interface
 \*********************************************************************************************/
 
-bool Xdrv82(uint8_t function) {
+bool Xdrv82(uint32_t function) {
   bool result = false;
 
   switch (function) {
