+++ tasmota/tasmota_xdrv_driver/xdrv_98_file_settings_demo.ino
@@ -174,7 +174,7 @@ void DrvDemoSettingsSave(void) {
  * Interface
 \*********************************************************************************************/
 
-bool Xdrv98(uint8_t function) {
+bool Xdrv98(uint32_t function) {
   bool result = false;
 
   switch (function) {
