+++ tasmota/tasmota_xdsp_display/xdsp_18_berry_display.ino
@@ -26,7 +26,7 @@
  * Interface
 \*********************************************************************************************/
 
-bool Xdsp18(uint8_t function) {
+bool Xdsp18(uint32_t function) {
   bool result = false;
 
   switch (function) {
