+++ tasmota/tasmota_xnrg_energy/xnrg_16_iem3000.ino
@@ -203,7 +203,7 @@ void Iem3000DrvInit(void)
  * Interface
 \*********************************************************************************************/
 
-bool Xnrg16(uint8_t function)
+bool Xnrg16(uint32_t function)
 {
   bool result = false;
